Correct the sampler shader to use uint formatted textures

+++ CapturePlugins/TanagerPlugin/ComputeShaders/Sampler_444_10bpc.hlsl
@@ -5,7 +5,7 @@ struct InputBufferStruct
 };
 
 StructuredBuffer<InputBufferStruct> InBuf : register(t0);
-RWTexture2D<float4> outputTexture : register(u0); // 16bpc uint 444 texture
+RWTexture2D<uint4> outputTexture : register(u0); // 16bpc uint 444 texture
 
 [numthreads(1, 1, 1)]
 void main(uint3 DTid : SV_DispatchThreadID)
